Move date navigation with calendar menu functionality into SimpleHeader
